requirements.txt
streamlit
transformers
qdrant-client
langchain
langchain-community
langchain-groq
sentence-transformers
logging
python-dotenv
langchain_qdrant
fastembed
